+++ doc/Changelog.md
@@ -3,9 +3,13 @@ Changelog {#Changelog}
 
 ## Deflect 0.8
 
-### 0.8.0 (21-09-2015)
+### 0.8.0 (git master)
 * Fix: ZeroConf record for server announces hostname instead of (wrong) port
 * Fix: send compressed JPEG images from correct thread, might fix crashes/hangs
+* Fix [#45](https://github.com/BlueBrain/Deflect/issues/45): Mouse position for
+  DesktopStreamer is rendered correctly on Retina displays
+* Fix interaction with DesktopStreamer on OSX where one click resulted in two
+  clicks
 * Added MTQueue class, moved from DisplayCluster
 * Added DeflectQml library to create QML applications which render offscreen
   and stream and receive events via Deflect